requirements-free.txt
# Zero-Cost Requirements (Perpetually Free)

# Web Framework & Async
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.25.0  # Compatible with python-telegram-bot
aiohttp>=3.9.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.12.0

# Database (FREE - SQLite built-in)
sqlalchemy>=2.0.0
alembic>=1.13.0

# Local LLM Inference (FREE)
ollama>=0.1.2  # Python client for Ollama
transformers>=4.36.0
torch>=2.1.0  # CPU version
sentence-transformers>=2.3.0

# Vector Database (FREE - Local)
chromadb>=0.4.22
lancedb>=0.4.0

# ML & Statistics (FREE)
scikit-learn>=1.4.0
xgboost>=2.0.0
optuna>=3.5.0
imbalanced-learn>=0.11.0

# Fine-Tuning (FREE - CPU/Colab)
peft>=0.7.0
accelerate>=0.25.0
datasets>=2.16.0
trl>=0.7.0

# Experiment Tracking (FREE - Local)
mlflow==2.9.2
tensorboard==2.15.1

# Web Scraping (FREE)
beautifulsoup4==4.12.3
scrapy==2.11.0
lxml==5.1.0

# Scheduler (FREE - GitHub Actions + Python)
schedule==1.2.0
apscheduler==3.10.4

# Reporting & Dashboard (FREE)
jinja2==3.1.3
markdown==3.5.2
plotly==5.18.0
streamlit==1.29.0

# Messaging (FREE)
python-telegram-bot==20.7

# Utilities
loguru==0.7.2
python-dotenv==1.0.0
tenacity==8.2.3
pydantic==2.5.3
pydantic-settings==2.1.0
pyyaml==6.0.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code Quality
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
